feat: make sure to reuse models
