+++ chainhook-subscriber/src/models/action.model.ts
@@ -1,0 +1,4 @@
+export interface Action {
+    type: string;
+    value: string;
+}
