+++ chainhook-subscriber/src/models/highest-bidder.model.ts
@@ -1,0 +1,6 @@
+import { Action } from "./action.model";
+
+export interface Highestbidder {
+    type: string;
+    value: Action;
+}
