+++ chainhook-subscriber/src/models/place-bid.model.ts
@@ -1,3 +1,6 @@
+import { Action } from "./action.model";
+import { Highestbidder } from "./highest-bidder.model";
+
 export interface PlaceBid {
     action: Action;
     'auction-id': Action;
@@ -13,14 +16,4 @@ export interface PlaceBid {
     'start-block': Action;
     'token-id': Action;
     type: Action;
-}
-
-export interface Highestbidder {
-    type: string;
-    value?: any;
-}
-
-export interface Action {
-    type: string;
-    value: string;
 }
