+++ chainhook-subscriber/src/models/whitelisted.model.ts
@@ -1,0 +1,4 @@
+export interface Whitelisted {
+    type: string;
+    value: boolean;
+}
